pixellab_integration/requirements.txt
pixellab>=1.0.5
pillow>=12.0.0
